+++ CHANGES.md
@@ -51,6 +51,7 @@
     is now used on the constant instead of the variable. This is needed
     in order to byteswap the constant at compile time in MSVC builds.
     (gcc does this even if the byteswap is on the variable.)
+  * Updated MiniZip to an unofficial fork: https://github.com/nmoinvaz/minizip
 
 ## v1.1 (released 2017/07/04)